+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## [1.2.0]
+### Added
  - `things register:scenario-set` Register (create or update) a scenario set for automatic communication under a product.
- - `things get:scenario-set` Get the scenario set registered under a product.
+ - `things get:scenario-set` Get the scenario set registered under a product
+ - `things remove:scenario-set` Delete a scenario set registered under a product
 
 ## [1.1.0]
 ### Added